requirements.txt
openai
gradio
python-dotenv
PyMuPDF
pypdf
Pillow
